split settings, colored logs, html logs

+++ Archiver/CameraArchiveHelper.py
@@ -1,12 +1,14 @@
-import json
-import os
-import shutil
+import json, os, shutil, logging
 from Archiver.FileArchive import FileArchive
 from Common.ElasticSearchHelper import ElasticSearchHelper
 from Archiver.CameraArchiveConfig import CameraArchiveConfig
 from Common.CommonHelper import CommonHelper
 
 class CameraArchiveHelper:
+
+    def __init__(self, logger: logging.Logger, isSimulation: bool = False):
+        self.log = logger
+        self.isSimulation = isSimulation
 
     def load_configs(self, dir, listConfig = []):
         configs = []
@@ -16,7 +18,7 @@ class CameraArchiveHelper:
             config = CameraArchiveConfig()
             config.fromJsonFile(os.path.join(dir, file))    
             if len(listConfig) == 0 or config.camera in listConfig:
-                print('Load Config: {} => {}'.format(file, config.camera))
+                self.log.info('Load Config: {} => {}'.format(file, config.camera))
                 configs.append(config)
         return configs
 
@@ -24,15 +26,15 @@ class CameraArchiveHelper:
         files = []
         for root, dirnames, filenames in os.walk(config.pathFrom()):
             if hasattr(config, 'ignore_dir') and self.dir_to_ignore(root, config.ignore_dir):
-                print('Ignore:', root)
+                self.log.info('Ignore:', root)
                 continue
             for filename in filenames: #fnmatch.filter(filenames, '*.c'):
                 try:
                     file = FileArchive(config, root, filename)
                     files.append(file)
                 except ValueError:
-                    print('FileArchive creation error')
-        print('{} files found'.format(len(files)))
+                    self.log.info('FileArchive creation error')
+        self.log.info(f'{len(files)} files found in {config.pathFrom()}')
         return files
 
     def dir_to_ignore(self, root: str, ignore_dir: []):
@@ -55,11 +57,11 @@ class CameraArchiveHelper:
             if last_file_dir_relative_to != dir_relative_to:
                 if last_file_dir_relative_to != '':
                     ext_stat = ' '.join(["{}:{}".format(k.upper(), exts[k]) for k in exts])
-                    print('\t{} files moved: {}. Total {}'.format(files_moved, ext_stat, common.size_human(bytes_moved)))
+                    self.log.info('\t{} files moved: {}. Total {}'.format(files_moved, ext_stat, common.size_human(bytes_moved)))
                     files_moved = 0
                     bytes_moved = 0
                     exts = {}
-                print('{} => {}'.format(dir_relative_frm, dir_relative_to))
+                self.log.info('{} => {}'.format(dir_relative_frm, dir_relative_to))
 
             files_moved += 1
             bytes_moved += file.frm.size()
@@ -69,9 +71,11 @@ class CameraArchiveHelper:
             os.makedirs(file.to.dir, exist_ok=True)
             ## os.rename("path/to/current/file.foo", "path/to/new/destination/for/file.foo")
             ##shutil.copy2(file.frm.path, file.to.path)
-            #print(file.frm.path, ' => ', file.to.path)
-            shutil.move(file.frm.path, file.to.path)
-            elastic.report_to_elastic(file)
+            if self.isSimulation:
+                self.log.info(f'{file.frm.path} => {file.to.path}')
+            else:
+                shutil.move(file.frm.path, file.to.path)
+                elastic.report_to_elastic(file)
 
             ext = file.to.get_extension()
             if ext not in exts:
@@ -82,7 +86,7 @@ class CameraArchiveHelper:
             #break
 
         ext_stat = ' '.join(["{}:{}".format(k.upper(), exts[k]) for k in exts])
-        print('\t{} files moved: {}. Total {}'.format(files_moved, ext_stat, common.size_human(bytes_moved)))
+        self.log.info('\t{} files moved: {}. Total {}'.format(files_moved, ext_stat, common.size_human(bytes_moved)))
         files_moved = 0
         bytes_moved = 0
-        print ('########################################################')
+        self.log.info('########################################################')
